Implement generic configuration system for Excel import validation
- Replace specific ImportColumnConfiguration with generic SystemConfiguration table
- Add flexible key-value configuration system supporting JSON, String, Number, Boolean types
- Implement SystemConfigurationRepository with generic type support
- Update ExcelProcessorService to use configurable column validation
- Add authentication system with JWT support (User entity, AuthService, UserService)
- Implement background job tracking system (BackgroundJob entity and services)
- Add comprehensive import validation and error handling
- Include migration files for all database schema changes
- Add brand management functionality
- Implement stock validation and batch processing services
- Update .gitignore to exclude large files and temporary office files

+++ InventorySystem.Application/DTOs/InventoryMovementDto.cs
@@ -8,10 +8,10 @@ public class InventoryMovementDto
     public DateTime Date { get; set; }
     public MovementType Type { get; set; }
     public string TypeDescription { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public string? Reason { get; set; }
-    public int PreviousStock { get; set; }
-    public int NewStock { get; set; }
+    public decimal PreviousStock { get; set; }
+    public decimal NewStock { get; set; }
     public string? DocumentNumber { get; set; }
     public string? UserName { get; set; }
     public string? Source { get; set; }
@@ -34,7 +34,7 @@ public class CreateInventoryMovementDto
 {
     public int ProductId { get; set; }
     public MovementType Type { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public string? Reason { get; set; }
     public string? DocumentNumber { get; set; }
     public string? UserName { get; set; }
@@ -58,7 +58,7 @@ public class ProductStockHistoryDto
     public int ProductId { get; set; }
     public string ProductCode { get; set; } = string.Empty;
     public string ProductName { get; set; } = string.Empty;
-    public int CurrentStock { get; set; }
+    public decimal CurrentStock { get; set; }
     public List<InventoryMovementDto> Movements { get; set; } = new();
-    public Dictionary<DateTime, int> StockByDate { get; set; } = new();
+    public Dictionary<DateTime, decimal> StockByDate { get; set; } = new();
 }
